refactor(orchestrator): Extract OptionsStrategyCoordinator and SessionManager
ArjanCodes clean architecture refactoring (Jan 10, 2026):

- Extract OptionsStrategyCoordinator (380 lines) from TradingOrchestrator
  - run_options_risk_check() - McMillan rules stop-loss/delta management
  - run_options_strategy() - Gate 6 Phil Town Rule #1 options
  - run_iv_options_execution() - Gate 7 IV-aware options pipeline

- Extract SessionManager (187 lines) from TradingOrchestrator
  - build_session_profile() - market day detection, ticker selection
  - maybe_reallocate_for_weekend() - weekend budget reallocation
  - is_us_market_day() - holiday-aware market day check

- TradingOrchestrator reduced from 3,234 to 2,854 lines (380 lines extracted)
- Each extracted class has single responsibility, easier to test/modify
- Delegations preserve backward compatibility - no API changes

Benefits:
- Options logic now isolated in one file for easier debugging
- Session management decoupled from trading logic
- Follows ArjanCodes' "act as architect" principle
